Modernize depaginator
This commit applies several modernizations to the depaginator.  First and
foremost, the Page type is discarded in favor of using generics syntax to
specify the type of items in the pages.  Second, the lists of errors are
condensed into a single error, utilizing the errors.Join functionality.
Finally, to avoid explosions of goroutines, the HandleItem method is now
called synchronously after retrieving a page.

Several other minor nits have been picked, including updating the Makefile
and addressing new lint complaints.