Decouple page getter from depaginator
While writing an application using depaginator, I noted that it was going
to be difficult to write tests for a page getter due to the GetPage
routine being passed the actual depaginator.  The issue is that we can't
easily create a depaginator object that can be used for testing,
disconnected from an API, so the user would have to make a whole fake API
just to test their page getter.

This change addresses this issue by abstracting the depaginator.  We
create a State interface with the three methods that may be called from a
PageGetter: Update, Request, and PerPage.  This then becomes the type that
is passed to the GetPage method.  This doesn't require any code changes,
as we can still pass the actual Depaginator object; but implementations of
PageGetter can now create a mock for the State object and test their
GetPage methods/functions in isolation.