Capture and report panics
This commit adds the necessary code to capture and report when the
goroutines panic.  A captured panic is reported like any other error,
with the error message including the object that was paniced and a stack
trace.  This only affects calls that are made within library goroutines;
if a panic occurs in the context of the caller thread, no effort is made
to capture that panic.  This means that the Starter and the Doner will not
be captured if they panic, but the PageGetter, the Handler, and the
Updater do have this handling.